refactor: gather filters and paths under scim2_models.path
Filters and paths share one grammar with two start rules, one syntax tree, one
resolver and one evaluator, yet the modules answering for them were split
between a filters package and five modules at the root of the distribution,
with the public namespace re-exporting mostly what lived outside it. Every
filter is built on attribute paths where a path may carry no filter, so the
attribute path is the notion the whole toolkit rests on, and the package is
named for it. scim2_models.filters is gone; from scim2_models.path import Path
keeps working.
