+++ tests/test_filter_semantics.py
@@ -21,22 +21,22 @@ from scim2_models import Schema
 from scim2_models import ScimFilter
 from scim2_models import SearchRequest
 from scim2_models import User
-from scim2_models.filters import AttrPath
-from scim2_models.filters import CompareOperator
-from scim2_models.filters import Comparison
-from scim2_models.filters import FilterVisitor
-from scim2_models.filters import LogicalExpr
-from scim2_models.filters import LogicalOperator
-from scim2_models.filters import Not
-from scim2_models.filters import Present
-from scim2_models.filters import ValuePath
-from scim2_models.filters import coerce_value
-from scim2_models.filters import compare
-from scim2_models.filters import is_present
-from scim2_models.filters import parse_filter
-from scim2_models.filters import resolve_attr_path
-from scim2_models.filters import resolve_comparison_path
-from scim2_models.grammar import _error_detail
+from scim2_models.path import AttrPath
+from scim2_models.path import CompareOperator
+from scim2_models.path import Comparison
+from scim2_models.path import FilterVisitor
+from scim2_models.path import LogicalExpr
+from scim2_models.path import LogicalOperator
+from scim2_models.path import Not
+from scim2_models.path import Present
+from scim2_models.path import ValuePath
+from scim2_models.path import coerce_value
+from scim2_models.path import compare
+from scim2_models.path import is_present
+from scim2_models.path import parse_filter
+from scim2_models.path import resolve_attr_path
+from scim2_models.path import resolve_comparison_path
+from scim2_models.path.grammar import _error_detail
 from scim2_models.resources.user import Email
 
 COMPOSED = "Jos\N{LATIN SMALL LETTER E WITH ACUTE}"
